+++ app/app.config.ts
@@ -3,9 +3,6 @@ export default defineAppConfig({
     colors: {
       primary: 'emerald',
       neutral: 'slate'
-    },
-    dashboardSidebarCollapse: {
-      base: 'flex'
     }
   }
 })